Add abstract type for wish list order item
Add abstract type for wish list order item
Add comments to order items classes
