+++ src/Demo/CartBundle/Entity/Cart/CartManager.php
@@ -10,7 +10,7 @@ namespace Demo\CartBundle\Entity\Cart;
 
 use Demo\CartBundle\Entity\OrderElement\OrderItem;
 use Demo\CartBundle\Entity\OrderElement\AbstractOrderItem;
-use Demo\CartBundle\Entity\OrderElement\WishOrderItemInterface;
+use Demo\CartBundle\Entity\OrderElement\DetailedWishItemInterface;
 use Demo\CartBundle\Entity\Product\AbstractShoppingItem;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareTrait;
@@ -110,7 +110,7 @@ class CartManager extends AbstractCartManager
     /**
      * @inheritDoc
      */
-    public function addWishOrderItem(WishOrderItemInterface $item)
+    public function addWishOrderItem(DetailedWishItemInterface $item)
     {
         // TODO: Implement addWishOrderItem() method.
     }
