+++ src/Demo/CartBundle/Entity/Cart/CartManagerInterface.php
@@ -8,7 +8,7 @@
 
 namespace Demo\CartBundle\Entity\Cart;
 use Demo\CartBundle\Entity\OrderElement\AbstractOrderItem;
-use Demo\CartBundle\Entity\OrderElement\WishOrderItemInterface;
+use Demo\CartBundle\Entity\OrderElement\DetailedWishItemInterface;
 use Demo\CartBundle\Entity\Product\AbstractShoppingItem;
 
 /**
@@ -77,10 +77,10 @@ interface CartManagerInterface
     /**
      * Add wish list item to the cart
      *
-     * @param WishOrderItemInterface $item
+     * @param DetailedWishItemInterface $item
      * @return mixed
      */
-    public function addWishOrderItem(WishOrderItemInterface $item);
+    public function addWishOrderItem(DetailedWishItemInterface $item);
 
     /**
      * remove item from the cart
