+++ src/Demo/CartBundle/Entity/Cart/ListManagerInterface.php
@@ -10,7 +10,7 @@ namespace Demo\CartBundle\Entity\Cart;
 
 
 use Demo\CartBundle\Entity\OrderElement\AbstractOrderItem;
-use Demo\CartBundle\Entity\OrderElement\WishOrderItemInterface;
+use Demo\CartBundle\Entity\OrderElement\DetailedWishItemInterface;
 use Demo\CartBundle\Entity\Product\AbstractShoppingItem;
 
 /**
@@ -78,10 +78,10 @@ interface ListManagerInterface
     /**
      * Add wish list item to the list
      *
-     * @param WishOrderItemInterface $item
+     * @param DetailedWishItemInterface $item
      * @return mixed
      */
-    public function addWishOrderItem(WishOrderItemInterface $item);
+    public function addWishOrderItem(DetailedWishItemInterface $item);
 
     /**
      * remove item from the list
