+++ src/Demo/CartBundle/Entity/OrderElement/ListItemInterface.php
@@ -11,6 +11,12 @@ namespace Demo\CartBundle\Entity\OrderElement;
 
 use Demo\CartBundle\Entity\Cart\AbstractWishList;
 
+/**
+ * Interface ListItemInterface
+ * @package Demo\CartBundle\Entity\OrderElement
+ *
+ * interface to define item that belongs to a certain list
+ */
 interface ListItemInterface
 {
     public function getList();
