+++ src/Demo/CartBundle/Entity/OrderElement/WishOrderItem.php
@@ -2,7 +2,6 @@
 
 namespace Demo\CartBundle\Entity\OrderElement;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -13,7 +12,7 @@ use Doctrine\ORM\Mapping as ORM;
  *
  * @ORM\Entity
  */
-class WishOrderItem extends ListOrderItem implements WishOrderItemInterface
+class WishOrderItem extends AbstractWishOrderItem implements DetailedWishItemInterface
 {
 
     /**
@@ -26,17 +25,6 @@ class WishOrderItem extends ListOrderItem implements WishOrderItemInterface
      */
     protected $important;
 
-    /**
-     * List to hold the sub items (sub orders fulfilled from the desired quantity)
-     * this is important to know that someone already bought this for you and who is he
-     *
-     *@ORM\ManyToMany(targetEntity="Demo\CartBundle\Entity\OrderElement\ListOrderItem")
-     * @ORM\JoinTable(name="wish_sub_items",
-     *      joinColumns={@ORM\JoinColumn(name="wish_item_id", referencedColumnName="id")},
-     *      inverseJoinColumns={@ORM\JoinColumn(name="sub_item_id", referencedColumnName="id", unique=true)}
-     *      )
-     */
-    protected $subItems;
 
     /**
      * WishOrderItem constructor.
@@ -45,8 +33,8 @@ class WishOrderItem extends ListOrderItem implements WishOrderItemInterface
      */
     public function __construct()
     {
+        parent::__construct();
         $this->important = false;
-        $this->subItems = new ArrayCollection();
     }
 
 
@@ -88,38 +76,6 @@ class WishOrderItem extends ListOrderItem implements WishOrderItemInterface
     public function getImportant()
     {
         return $this->important;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function addSubItem(ListOrderItem $subItem)
-    {
-        // TODO: Implement addSubItem() method.
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getSubItems()
-    {
-        // TODO: Implement getSubItems() method.
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function remainingCount()
-    {
-        // TODO: Implement remainingCount() method.
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function isFulfilled()
-    {
-        // TODO: Implement isFulfilled() method.
     }
